% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avaliacao-utils.R
\name{questao}
\alias{questao}
\title{Questão}
\usage{
questao(
  text,
  ...,
  type = c("auto", "single", "multiple", "learnr_radio", "learnr_checkbox",
    "learnr_text")[1]
)
}
\arguments{
\item{text}{Título da questão}

\item{...}{answer()'s e outros arguemntos para \link[learnr]{question}}

\item{type}{tipo da questão}
}
\value{

}
\description{
Versão do learnr::question() em português
}
